refactor(tags): the merge moves to its own file
tagvocab.go crossed the 500-line cap with the survivor's pin. A merge is
the one vocabulary act that writes two tag rows and every taggable row
between them, and it carries the preconditions that go with that — so it is
a concept of its own rather than another entry beside create and rename.
